Add frame latency control
Set frame latency to 2, which means the number outstanding inflight GPU
submissions is 2. The host wait operation is put after the vkQueuePresentKHR.
However, the swapchain(both EGL and Vulkan) behavior on Android grants this
latency control somehow by blocking in queueBuffer(vkQueuePresentKHR and
eglSwapBuffers) to not allow outstanding GPU works to be more than 2. But who
knows when the black magic would go away.