Make sure pre-rotation is triggered on split screen
On split screen, rotating 90 degree won't trigger onNativeWindowResized callback
sometimes. So I set a countdown to guarantee a swapchain recreation upon seeing
VK_SUBOPTIMAL_KHR for kPreRotationLatency times.

Although APP_CMD_CONFIG_CHANGED is still triggered in the above case, I can't
rely on that as the signal to recreate the swapchain because at that time, the
surface size hasn't been updated to the new dimension yet, which is so annoying.